requirements.txt
numpy==1.24.3
scipy==1.11.3
Pillow==10.0.0
watchdog==3.0.0 